+++ traffic_router/core/src/main/java/org/apache/traffic_control/traffic_router/core/dns/protocol/TCP.java
@@ -46,9 +46,9 @@ public class TCP extends AbstractProtocol {
     @Override
     public void run() {
         while (!isShutdownRequested()) {
+            final TCPSocketHandler handler;
             try {
-                final Socket socket = getServerSocket().accept();
-                final TCPSocketHandler handler = new TCPSocketHandler(socket);
+                handler = new TCPSocketHandler(getServerSocket().accept());
                 submit(handler);
             } catch (final IOException e) {
 				LOGGER.warn("error: " + e);
@@ -95,13 +95,13 @@ public class TCP extends AbstractProtocol {
                 return;
             }
 
-            try {
+            try (InputStream iis = Channels.newInputStream(Channels.newChannel(socket.getInputStream()));
+                 DataInputStream is = new DataInputStream(iis);
+                 DataOutputStream os = new DataOutputStream(socket.getOutputStream())
+            ) {
                 socket.setSoTimeout(getReadTimeout());
                 final InetAddress client = socket.getInetAddress();
 
-                final InputStream iis = Channels.newInputStream(Channels.newChannel(socket.getInputStream()));
-                final DataInputStream is = new DataInputStream(iis);
-                final DataOutputStream os = new DataOutputStream(socket.getOutputStream());
                 final int length = is.readUnsignedShort();
                 final byte[] request = new byte[length];
                 is.readFully(request);